Add --enable-jemalloc flag to build system
The OVS build system supports the enable-jemalloc flag to generate
the ovsdb-server binary using jemalloc. This flag, combined with
the change in the Yocto recipe, allows to switch the memory manager
used by the database running in the image.

Also contains a markdown that describes the change, and memory usage and
time improvements.

Change-Id: I81493028a681beddaeb525102021d6ce32801c78
Signed-off-by: Esteban Rodriguez Betancourt <estebarb@hpe.com>