Support for on-demand fetching of non-monitored data
The IDL only support reading information from data that is part of the
replica. For processes that rarely access the data base, or part of it,
this could result in a waste of resources.

This commit adds support for reading data that is not being monitored
from the OVSDB server. The data can be retrived at row, column, table
level.

Change-Id: Idedca4263ca932a5d9d3143133ad4d020a1c6baa
Signed-off-by: Sebastian Arguello <sebastian.arguello@hpe.com>