new: dev: on-demand fetching of non-monitored data
The IDL only support reading information from data that is part of the
replica. For processes that rarely access the data base, or part of it,
this could result in a waste of resources.

This commit adds support for reading data that is not being monitored
from the OVSDB server. The data can be retrived at row, column, table
level. There are function generated for each table.

Change-Id: Ide693ef364ab5ea1dce373042331006914487470
Signed-off-by: Sebastian Arguello <sebastian.arguello@hpe.com>