vswitchd changes to support sflow per interface.
vSwitchd changes to support sflow on each interface. Per interface sflow
config is saved in 'other_config' column of a port as a key:value pair. The
key is: "sflow-enable" string and value is either "true" or "false" if sflow
is _explicitly_ enabled or disabled respectively.

Default value on each port is "true".

Change-Id: I27f4a196ab910152b0a9191580ae0f47e9820ed1
Signed-off-by: Bhaskar Upadhyayula <sarma.bha.upadhyayula@hpe.com>