chg: dev: Add PI/PD API for ACL logging
This commit adds the sequence number that the PD code will increment to
wake up the main loop in switchd when an ACL logging packet is received.
The acllog_info_t struct includes the data members that PD can pass to
PI when a packet is received.  If the PD code cannot identify the ACE
that the packet matched, the valid_fields member will indicate that the
ACE-related fields are not valid.

Change-Id: I67096e6818aa5531ac53f1e9f106a7e520441022
Signed-off-by: Gary Wassermann <gary.wassermann@hpe.com>