new: dev: OVSDB Priority Sessions
The priority sessions feature adds to the OVSDB Server
the ability to identify and prioritize client sessions.

The clients identify themselves using a new JSON RPC command,
"identify", which receives the daemon name as parameter or the
client's identifier.

Then the server compares that identifier with a list of
priorities assigned by name. If the name matches, then that
priority is selected, otherwise a default priority is assigned.

The priorities are saved in a json file, that the ovsdb server
receives as a optional parameter when initializing. With
ovsdb-appctl the ovsdb-server can be forced to reload that file
in runtime.

The priorities goes from 0 to 15, being 0 the most important,
and 15 the less important.

The OVSDB Server process the requests in a loop. The priority
system works deferring the processing of the requests with a
lower priority to a next iteration of the loop.

The sessions with priority 0 run all the times (16 times of each
16 iterations), with priority 1 run 15 times of each 16 iterations,
and so on, finally with priority 15, that runs just 1 time of each
16 iterations.

The JSON RPC was also added to the ovsdb-client program.

Change-Id: I09f36243a19e27c2f2a179d6a55a8a07211f71a9
Signed-off-by: Esteban Rodriguez Betancourt <estebarb@hpe.com>