Enabling Open vSwitch tests
This changes enable the compiling of Open vSwitch tests inside
OPS-BUILD.
The tests related to Open vSwitch modules that aren't compiled in
ops-build are excluded from the code generation and some tests that
depend on ovs-vswitchd are skipped. The files remain in the folder.

Check file  lib/socket-util-unix.c for a pending condition that need to be
defined to run the tests in native host.

Change-Id: Ide695f9bbe801e3f51c3c094df8fe32f8e32886e
Signed-off-by: arnoldo.lutz.guevara@hpe.com <arnoldo.lutz.guevara@hpe.com>