ovsdb: provide better context inside log messages from ovsdb server when reporting trasaction errors.
TG-669

When OVSDB reports errors, the messages should give enough information so that the problem can be triaged.
When large transactions are attempted and a constraint violation is flagged on one of the elements, the log
message should also identify the failing Table & Column.

Right now, the log message only contains the failing constraint, but does not identify which Table or Column
was being operated on when the failure occured.

Change-Id: Ie116f471059b1b7699b4690620957c35be4f72a4
Signed-off-by: Surya Tatapudi <surya.tatapudi@hpe.com>